Extract Potassium Slip rendering into helper
